# Improve RegexOptions.Compiled codegen for ignoring case (in particular InvariantCulture)