Improved RegexOptions.Compiled code gen, in particular for IgnoreCase

Several changes:
- Improve the code generation for finding the first character in FindFirstChar, using a pattern the JIT can better optimize.
- Avoid using ToLower{Invariant} if we’re going to codegen out a call to a method that’s already case-insensitive, like char.IsDigit/IsWhiteSpace, and when using InvariantCulture, build the case-insensitivity into the lookup table so that we can similarly avoid the ToLower{Invariant} call on the fast path.
- Make the timeout check faster by using the unsigned rem.un instead of the signed rem
- Put back a couple of writes I accidentally removed in a previous PR, resulting in duplicating some unnecessary work in some cases.

cc: @danmosemsft, @eerhardt, @ViktorHofer, @GrabYourPitchforks 

|    Method |        Toolchain |              Options |       Mean |    Error | Ratio |
|---------- |----------------- |--------------------- |-----------:|---------:|------:|
|     Email | \new\corerun.exe | Ignor(...)riant [38] |   172.0 ns |  1.38 ns |  0.62 |
|     Email | \old\corerun.exe | Ignor(...)riant [38] |   275.8 ns |  4.14 ns |  1.00 |
|           |                  |                      |            |          |       |
|      Date | \new\corerun.exe | Ignor(...)riant [38] |   109.8 ns |  0.47 ns |  0.60 |
|      Date | \old\corerun.exe | Ignor(...)riant [38] |   182.8 ns |  0.97 ns |  1.00 |
|           |                  |                      |            |          |       |
|        IP | \new\corerun.exe | Ignor(...)riant [38] |   647.4 ns |  2.66 ns |  0.70 |
|        IP | \old\corerun.exe | Ignor(...)riant [38] |   930.4 ns |  5.17 ns |  1.00 |
|           |                  |                      |            |          |       |
|       Uri | \new\corerun.exe | Ignor(...)riant [38] |   110.7 ns |  0.85 ns |  0.66 |
|       Uri | \old\corerun.exe | Ignor(...)riant [38] |   168.8 ns |  1.29 ns |  1.00 |
|           |                  |                      |            |          |       |
| UriSpaces | \new\corerun.exe | Ignor(...)riant [38] |   300.4 ns |  2.27 ns |  0.34 |
| UriSpaces | \old\corerun.exe | Ignor(...)riant [38] |   886.7 ns |  6.39 ns |  1.00 |
|     Email | \new\corerun.exe | IgnoreCase, Compiled |   307.8 ns |  7.13 ns |  0.99 |
|     Email | \old\corerun.exe | IgnoreCase, Compiled |   312.6 ns |  3.55 ns |  1.00 |
|           |                  |                      |            |          |       |
|      Date | \new\corerun.exe | IgnoreCase, Compiled |   119.9 ns |  0.40 ns |  0.61 |
|      Date | \old\corerun.exe | IgnoreCase, Compiled |   196.7 ns |  1.52 ns |  1.00 |
|           |                  |                      |            |          |       |
|        IP | \new\corerun.exe | IgnoreCase, Compiled | 1,097.5 ns | 16.10 ns |  1.03 |
|        IP | \old\corerun.exe | IgnoreCase, Compiled | 1,067.8 ns |  8.51 ns |  1.00 |
|           |                  |                      |            |          |       |
|       Uri | \new\corerun.exe | IgnoreCase, Compiled |   193.1 ns |  1.07 ns |  0.99 |
|       Uri | \old\corerun.exe | IgnoreCase, Compiled |   194.5 ns |  1.85 ns |  1.00 |
|           |                  |                      |            |          |       |
| UriSpaces | \new\corerun.exe | IgnoreCase, Compiled |   966.6 ns | 11.87 ns |  1.00 |
| UriSpaces | \old\corerun.exe | IgnoreCase, Compiled |   966.2 ns |  4.03 ns |  1.00 |
|           |                  |                      |            |          |       |

```C#
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;
using System.Text.RegularExpressions;

public class Program
{
    static void Main(string[] args) => BenchmarkSwitcher.FromAssemblies(new[] { typeof(Program).Assembly }).Run(args);
}

public class Regexes
{
    [Params(
        RegexOptions.Compiled | RegexOptions.IgnoreCase,
        RegexOptions.Compiled | RegexOptions.IgnoreCase | RegexOptions.CultureInvariant)]
    public RegexOptions Options { get; set; }

    private Regex _email, _date, _ip, _uri;

    private const string EmailPattern = @"^([a-zA-Z0-9_\-\.]+)@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.)|(([a-zA-Z0-9\-]+\.)+))([a-zA-Z]{2,12}|[0-9]{1,3})(\]?)$";
    private const string DatePattern = @"\b(?<month>\d{1,2})/(?<day>\d{1,2})/(?<year>\d{2,4})\b";
    private const string IPPattern = @"(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9])\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9])";
    private const string UriPattern = @"[\w]+://[^/\s?#]+[^\s?#]+(?:\?[^\s#]*)?(?:#[^\s]*)?";
    private const string LotsOfSpaces = "                                                                                                                                                          ";

    [GlobalSetup]
    public void Setup()
    {
        _email = new Regex(EmailPattern, Options);
        _date = new Regex(DatePattern, Options);
        _ip = new Regex(IPPattern, Options);
        _uri = new Regex(UriPattern, Options);
    }

    [Benchmark] public void Email() => _email.IsMatch("yay.performance@dot.net");
    [Benchmark] public void Date() => _date.IsMatch("Today is 11/18/2019");
    [Benchmark] public void IP() => _ip.IsMatch("012.345.678.910");
    [Benchmark] public void Uri() => _uri.IsMatch("http://example.org");
    [Benchmark] public void UriSpaces() => _uri.IsMatch(LotsOfSpaces + "http://example.org");
}
```