@stephentoub 2019-12-11
@danmosemsft, I pushed a style-only commit to address the feedback (mostly changing things I didn't actually add in this PR).